Backend/requirements.txt
google-genai
langchain
langchain-community
langchain-chroma
chromadb
transformers
torch
pydantic
python-dotenv
sentence-transformers